requirements.txt
# Python 3.10+ (testowane z 3.13)
pandas>=2.0
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
Pillow>=10.0
torch>=2.0
torchvision>=0.15
tqdm>=4.65
pyarrow>=12.0
